requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
pytest>=7.0.0
pydantic>=2.0.0 